Renamed InvalidLogin.ts to InvalidLoginError.ts
